remove all code releated to react
