+++ package.json
@@ -33,22 +33,11 @@
     "@nhost/hasura-storage-js": "^0.0.3",
     "axios": "^0.21.1",
     "jwt-decode": "^3.1.2",
-    "query-string": "^7.0.1",
-    "subscriptions-transport-ws": "^0.9.19"
+    "query-string": "^7.0.1"
   },
   "devDependencies": {
-    "@apollo/client": "^3.4.0",
-    "@types/react": "^17.0.17",
-    "@types/ws": "^7.4.7",
-    "graphql": "^15.5.1",
     "np": "^7.5.0",
-    "react": "^16.0.0 || ^17.0.0",
-    "react-dom": "^16.0.0 || ^17.0.0",
     "typescript": "^4.4.2"
   },
-  "peerDependencies": {
-    "@apollo/client": "^3.4.0",
-    "react": "^16.0.0 || ^17.0.0",
-    "react-dom": "^16.0.0 || ^17.0.0"
-  }
+  "peerDependencies": {}
 }
