--- src/react/context/index.ts
@@ -1,2 +1,0 @@
-export * from './nhost';
-export * from './apollo';
