--- src/react/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from './useNhost';
-export * from './useNhostAuth';
