--- src/react/hooks/useNhost.ts
@@ -1,7 +1,0 @@
-import { useContext } from 'react';
-import { NhostClientContext } from '../context';
-
-export function useNhost() {
-  const nhost = useContext(NhostClientContext);
-  return { nhost: nhost! };
-}
