--- src/react/index.ts
@@ -1,2 +1,0 @@
-export * from './context';
-export * from './hooks';
